+++ CHANGELOG.rst
@@ -23,6 +23,14 @@ Version 0.9 (unreleased)
   function now support geometries with z-coordinates (#131).
 * Addition of Cython and internal PyGEOS C API to enable easier development of internal
   functions (previously all significant internal functions were developed in C) (#51).
+* API change: geometry and counting functions (``get_num_coordinates``,
+  ``get_num_geometries``, ``get_num_interior_rings``, ``get_num_points``) now return 0
+  for ``None`` input values instead of -1 (#218).
+* Fixed internal GEOS error code detection for ``get_dimensions`` and ``get_srid`` (#218).
+* Addition of ``prepare`` function that generates a GEOS prepared geometry which is stored on
+  the Geometry object itself. All binary predicates (except ``equals``) make use of this (#92).
+* Use previously prepared geometries within ``STRtree`` ``query`` and ``query_bulk``
+  functions if available (#246).
 * Ensure that ``python setup.py clean`` removes all previously Cythonized and compiled
   files (#239).